// backend/src/module/common/common.type.ts
export interface User {
  id: string;
  email: string;
  password_hash: string;
  full_name: string;
  username: string;
  avatar_url: string | null;
  bio: string | null;
  school: string | null;
  major: string | null;
  role: 'user' | 'admin';
  is_banned: boolean;
  created_at: Date;
  updated_at: Date;
}
// ...
export interface TokenPayload {
  userId: string;
  role: string;
}

export interface AuthTokens {
  accessToken: string;
  refreshToken: string;
}
// ...
export interface UserPublic {
  id: string;
  email: string;
  full_name: string;
  username: string;
  avatar_url: string | null;
  bio: string | null;
  school: string | null;
  major: string | null;
  role: string;
  created_at: Date;
}